requirements.txt
bilibili-api-python>=17.0.0
aiohttp>=3.11.0
fastapi>=0.110.0
uvicorn>=0.23.0
python-dotenv>=1.0.0
